requirements.txt
langchain
langchain-groq
psycopg2-binary==2.9.7
langgraph
python-dotenv
groq
